+++ small_text/classifiers/__init__.py
@@ -1,6 +1,20 @@
 from small_text.classifiers.classification import (
     Classifier,
     SklearnClassifier,
+    EmbeddingMixin,
     ConfidenceEnhancedLinearSVC
 )
-from small_text.classifiers.factories import SklearnClassifierFactory
+from small_text.classifiers.factories import (
+    AbstractClassifierFactory,
+    SklearnClassifierFactory
+)
+
+
+__all__ = [
+    'Classifier',
+    'SklearnClassifier',
+    'EmbeddingMixin',
+    'ConfidenceEnhancedLinearSVC',
+    'AbstractClassifierFactory',
+    'SklearnClassifierFactory'
+]