+++ small_text/initialization/__init__.py
@@ -3,3 +3,10 @@ from small_text.initialization.strategies import (
     random_initialization_balanced,
     random_initialization_stratified
 )
+
+
+__all__ = [
+    'random_initialization',
+    'random_initialization_balanced',
+    'random_initialization_stratified'
+]