+++ small_text/integrations/pytorch/__init__.py
@@ -1,8 +1,44 @@
 from small_text.integrations.pytorch.exceptions import PytorchNotFoundError
 
 try:
-    from small_text.integrations.pytorch.datasets import (
-        PytorchTextClassificationDataset
+    from small_text.integrations.pytorch.classifiers.base import (
+        PytorchModelSelectionMixin,
+        PytorchClassifier
     )
+    from small_text.integrations.pytorch.classifiers.kimcnn import (
+        kimcnn_collate_fn,
+        KimCNNEmbeddingMixin,
+        KimCNNClassifier
+    )
+    from small_text.integrations.pytorch.datasets import (
+        PytorchDataset,
+        PytorchDatasetView,
+        PytorchTextClassificationDataset,
+        PytorchTextClassificationDatasetView
+    )
+    from small_text.integrations.pytorch.models.kimcnn import KimCNN
+    from small_text.integrations.pytorch.query_strategies.strategies import (
+        ExpectedGradientLength,
+        ExpectedGradientLengthLayer,
+        ExpectedGradientLengthMaxWord
+    )
+
+    __all__ = [
+        'PytorchModelSelectionMixin',
+        'PytorchClassifier',
+        'KimCNN',
+        'kimcnn_collate_fn',
+        'KimCNNEmbeddingMixin',
+        'KimCNNClassifier',
+        'ExpectedGradientLength',
+        'ExpectedGradientLengthLayer',
+        'ExpectedGradientLengthMaxWord',
+        'PytorchNotFoundError',
+        'PytorchDataset',
+        'PytorchDatasetView',
+        'PytorchTextClassificationDataset',
+        'PytorchTextClassificationDatasetView'
+    ]
+
 except PytorchNotFoundError:
-    pass
+    __all__ = []