+++ small_text/integrations/pytorch/classifiers/__init__.py
@@ -1,2 +1,18 @@
-from small_text.integrations.pytorch.classifiers.base import PytorchClassifier
+from small_text.integrations.pytorch.classifiers.base import (
+    PytorchClassifier,
+    PytorchModelSelectionMixin
+)
 from small_text.integrations.pytorch.classifiers.factories import KimCNNFactory
+from small_text.integrations.pytorch.classifiers.kimcnn import (
+    KimCNNEmbeddingMixin,
+    KimCNNClassifier
+)
+
+
+__all__ = [
+    'PytorchClassifier',
+    'PytorchModelSelectionMixin',
+    'KimCNNFactory',
+    'KimCNNEmbeddingMixin',
+    'KimCNNClassifier'
+]