+++ small_text/integrations/pytorch/models/__init__.py
@@ -1,0 +1,5 @@
+from small_text.integrations.pytorch.models.kimcnn import KimCNN
+
+__all__ = [
+    'KimCNN'
+]