+++ small_text/integrations/pytorch/query_strategies/__init__.py
@@ -4,3 +4,11 @@ from small_text.integrations.pytorch.query_strategies.strategies import (
     ExpectedGradientLengthLayer,
     ExpectedGradientLengthMaxWord
 )
+
+
+__all__ = [
+    'BADGE',
+    'ExpectedGradientLength',
+    'ExpectedGradientLengthLayer',
+    'ExpectedGradientLengthMaxWord'
+]