+++ small_text/integrations/pytorch/utils/__init__.py
@@ -1,0 +1,7 @@
+from small_text.integrations.pytorch.utils.data import dataloader, get_class_weights
+
+
+__all__ = [
+    'dataloader',
+    'get_class_weights'
+]