+++ small_text/utils/__init__.py
@@ -1,0 +1,40 @@
+from small_text.utils.annotations import DeprecationError, ExperimentalWarning
+from small_text.utils.classification import split_data, prediction_result, empty_result
+from small_text.utils.clustering import init_kmeans_plusplus_safe
+from small_text.utils.context import build_pbar_context, NullProgressBar
+from small_text.utils.data import check_training_data, list_length
+from small_text.utils.labels import get_num_labels, csr_to_list, list_to_csr
+from small_text.utils.logging import (
+    verbosity_logger,
+    VerbosityLogger,
+    VERBOSITY_QUIET,
+    VERBOSITY_VERBOSE,
+    VERBOSITY_MORE_VERBOSE,
+    VERBOSITY_ALL
+)
+from small_text.utils.system import get_tmp_dir_base, TMP_DIR_VARIABLE
+
+
+__all__ = [
+    'DeprecationError',
+    'ExperimentalWarning',
+    'split_data',
+    'empty_result',
+    'init_kmeans_plusplus_safe',
+    'build_pbar_context',
+    'NullProgressBar',
+    'build_pbar_context',
+    'check_training_data',
+    'list_length',
+    'get_num_labels',
+    'csr_to_list',
+    'list_to_csr',
+    'verbosity_logger',
+    'VerbosityLogger',
+    'VERBOSITY_QUIET',
+    'VERBOSITY_VERBOSE',
+    'VERBOSITY_MORE_VERBOSE',
+    'VERBOSITY_ALL',
+    'get_tmp_dir_base',
+    'TMP_DIR_VARIABLE'
+]