+++ tests/unit/small_text/integrations/transformers/test_imports.py
@@ -1,0 +1,17 @@
+import unittest
+import pytest
+
+
+@pytest.mark.pytorch
+class ImportsTest(unittest.TestCase):
+
+    def test_import_datasets_modules(self):
+        from small_text import TransformersDataset
+        from small_text import TransformersDatasetView
+
+    def test_import_classifiers_modules(self):
+        from small_text import transformers_collate_fn
+        from small_text import FineTuningArguments
+        from small_text import TransformerModelArguments
+        from small_text import TransformerBasedEmbeddingMixin
+        from small_text import TransformerBasedClassification